Remove redundant tests from core/cc package
The integration style tests in the chaincode package take a very long
time to run as they compile and run chaincode in docker containers.
Given we already have project level integration tests that exercise this
code, we can remove these without taking on any risk.
