--- core/chaincode/testdata/src/chaincodes/example02/chaincode_test.go
@@ -1,104 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package main
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/hyperledger/fabric-chaincode-go/shim"
-	"github.com/hyperledger/fabric-chaincode-go/shimtest"
-)
-
-func checkInit(t *testing.T, stub *shimtest.MockStub, args [][]byte) {
-	res := stub.MockInit("1", args)
-	if res.Status != shim.OK {
-		fmt.Println("Init failed", string(res.Message))
-		t.FailNow()
-	}
-}
-
-func checkState(t *testing.T, stub *shimtest.MockStub, name string, value string) {
-	bytes := stub.State[name]
-	if bytes == nil {
-		fmt.Println("State", name, "failed to get value")
-		t.FailNow()
-	}
-	if string(bytes) != value {
-		fmt.Println("State value", name, "was not", value, "as expected")
-		t.FailNow()
-	}
-}
-
-func checkQuery(t *testing.T, stub *shimtest.MockStub, name string, value string) {
-	res := stub.MockInvoke("1", [][]byte{[]byte("query"), []byte(name)})
-	if res.Status != shim.OK {
-		fmt.Println("Query", name, "failed", string(res.Message))
-		t.FailNow()
-	}
-	if res.Payload == nil {
-		fmt.Println("Query", name, "failed to get value")
-		t.FailNow()
-	}
-	if string(res.Payload) != value {
-		fmt.Println("Query value", name, "was not", value, "as expected")
-		t.FailNow()
-	}
-}
-
-func checkInvoke(t *testing.T, stub *shimtest.MockStub, args [][]byte) {
-	res := stub.MockInvoke("1", args)
-	if res.Status != shim.OK {
-		fmt.Println("Invoke", args, "failed", string(res.Message))
-		t.FailNow()
-	}
-}
-
-func TestExample02_Init(t *testing.T) {
-	scc := new(SimpleChaincode)
-	stub := shimtest.NewMockStub("ex02", scc)
-
-	// Init A=123 B=234
-	checkInit(t, stub, [][]byte{[]byte("init"), []byte("A"), []byte("123"), []byte("B"), []byte("234")})
-
-	checkState(t, stub, "A", "123")
-	checkState(t, stub, "B", "234")
-}
-
-func TestExample02_Query(t *testing.T) {
-	scc := new(SimpleChaincode)
-	stub := shimtest.NewMockStub("ex02", scc)
-
-	// Init A=345 B=456
-	checkInit(t, stub, [][]byte{[]byte("init"), []byte("A"), []byte("345"), []byte("B"), []byte("456")})
-
-	// Query A
-	checkQuery(t, stub, "A", "345")
-
-	// Query B
-	checkQuery(t, stub, "B", "456")
-}
-
-func TestExample02_Invoke(t *testing.T) {
-	scc := new(SimpleChaincode)
-	stub := shimtest.NewMockStub("ex02", scc)
-
-	// Init A=567 B=678
-	checkInit(t, stub, [][]byte{[]byte("init"), []byte("A"), []byte("567"), []byte("B"), []byte("678")})
-
-	// Invoke A->B for 123
-	checkInvoke(t, stub, [][]byte{[]byte("invoke"), []byte("A"), []byte("B"), []byte("123")})
-	checkQuery(t, stub, "A", "444")
-	checkQuery(t, stub, "B", "801")
-
-	// Invoke B->A for 234
-	checkInvoke(t, stub, [][]byte{[]byte("invoke"), []byte("B"), []byte("A"), []byte("234")})
-	checkQuery(t, stub, "A", "678")
-	checkQuery(t, stub, "B", "567")
-	checkQuery(t, stub, "A", "678")
-	checkQuery(t, stub, "B", "567")
-}
